ai-healer-django/requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
Django==5.0
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-import-export==4.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
import-export==0.3.1
mysqlclient==2.2.7
pillow==12.1.0
PyJWT==2.10.1
python-slugify==8.0.4
sqlparse==0.5.5
tablib==3.9.0
text-unidecode==1.3
typing_extensions==4.15.0

# DOM Extraction and Matching Engine dependencies
beautifulsoup4==4.13.4
numpy==2.2.2
pandas==2.2.3
scikit-learn==1.6.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
